* commit '7d7f3d9aa0e159bba21e95787bd69a444b409d7f':
  Add memory layout enum to tensor function signatures [PSD-807].
  This will allow support for nested and contiguous memory layouts.
  Enum values for fortran layouts have been included, but Fortran
  functionality does not yet exist. Improve the parsing of the
  environment variable SSDB for determining the database
  nodes to connect to [PSD-772]. Unify the get_tensor()
  and unpack_tensor() methods in SmartSimClient and DataSet
  objects [PSD-774]. C++ tests have been updated to test both
  get_tensor and unpack_tensor.  Only 1D tensor put, get,
  and unpack tensors have been updated for the C client
  because of future work on memory layout options that
  will continue to change the C tests and C API.

+++ src/cpp/commandreply.cpp
@@ -162,7 +162,7 @@ int CommandReply::has_error()
     if(this->_reply->type == REDIS_REPLY_ERROR)
         num_errors++;
     else if(this->_reply->type == REDIS_REPLY_ARRAY) {
-        for(int i=0; i<this->_reply->elements; i++) {
+        for(size_t i=0; i<this->_reply->elements; i++) {
           CommandReply tmp = (*this)[i];
           num_errors += tmp.has_error();
         }
@@ -181,7 +181,7 @@ void CommandReply::print_reply_error()
         std::cout<<error<<std::endl;
     }
     else if(this->_reply->type == REDIS_REPLY_ARRAY) {
-        for(int i=0; i<this->_reply->elements; i++) {
+        for(size_t i=0; i<this->_reply->elements; i++) {
           CommandReply tmp = (*this)[i];
           tmp.print_reply_error();
         }
@@ -257,7 +257,7 @@ void CommandReply::print_reply_structure(std::string index_tracker)
                <<std::endl;
       break;
     case REDIS_REPLY_ARRAY:
-      for(int i=0; i<this->n_elements(); i++) {
+      for(size_t i=0; i<this->n_elements(); i++) {
         std::string r_prefix = index_tracker + "[" + std::to_string(i) + "]";
         CommandReply tmp = (*this)[i];
         tmp.print_reply_structure(r_prefix);
